pyproject.toml
# SPDX-FileCopyrightText: 2024 Timeseer.AI
# SPDX-License-Identifier: Apache-2.0

[project]
name = "kukur"
version = "1.21.0"
description = "Kukur makes time series data and metadata available to the Apache Arrow ecosystem."
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: Apache Software License",
]
authors = [{ name = "Timeseer.AI", email = "pypi@timeseer.ai" }]
requires-python = ">=3.10"
dependencies = [
    "pyarrow>=16.0.0",
    "python-dateutil>=2.8.1",
    "toml>=0.10.2; python_version < '3.11'",
    "pytz>=2021.1",
]

[project.license]
text = "Apache-2.0"

[project.urls]
Homepage = "https://kukur.timeseer.ai/"
Documentation = "https://kukur.timeseer.ai/"
Repository = "https://github.com/timeseer-ai/kukur"
Issues = "https://github.com/timeseer-ai/kukur/issues"

[project.optional-dependencies]
adodb = ["pywin32; sys_platform == 'win32'"]
cratedb = ["crate"]
datafusion = ["datafusion"]
delta = ["deltalake"]
excel = ["openpyxl"]
influxdb = ["influxdb"]
kusto = ["azure-identity", "azure-kusto-data"]
odbc = ["pyodbc"]
numpy = ["numpy"]
piwebapi = ["requests", "requests-kerberos"]
postgresql = ["psycopg[binary]", "pg8000"]
redshift = ["redshift_connector"]
simulator = ["PyYAML"]

[dependency-groups]
dev = [
    "black >=24.4.0",
    "cx_Freeze >=7.2.7; sys_platform == 'win32'",
    "cx-Logging >=3.0; sys_platform == 'win32'",
    "mypy >=1.9.0,< 2",
    "pytest >=8.1.1,< 9",
    "ruff >=0.8.3,< 0.10",
    "types-python-dateutil",
    "types-pytz",
    "types-PyYAML",
    "types-requests",
    "types-toml; python_version < '3.11'",
]

[project.scripts]
kukur = "kukur.cli:_run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["kukur"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = [
    "ARG",
    "B",
    "C4",
    "D",
    "DTZ",
    "E",
    "F",
    "I",
    "N",
    "PD",
    "PL",
    "PLW",
    "Q",
    "RET",
    "T20",
    "UP",
]
ignore = ["C408", "D105", "D107", "D203", "D213", "RET504"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"kukur/__init__.py" = ["I001"]
"tests/**.py" = ["ARG001", "ARG002", "D", "PLR"]

[tool.pytest.ini_options]
markers = ["crate", "elasticsearch", "influxdb", "kukur", "odbc", "postgresql"]

[[tool.cxfreeze.executables]]
script = "kukur/cli.py"
target_name = "kukur.exe"

[[tool.cxfreeze.executables]]
script = "windows/win_service_config.py"
base = "Win32Service"
target_name = "kukur-service.exe"

[tool.cxfreeze.build_exe]
include_msvcr = true
includes = ["cx_Logging", "windows.win_service"]
include_files = [["windows/Kukur-windows.toml", "Kukur-example.toml"]]

[tool.cxfreeze.bdist_msi]
upgrade_code = "{859ffb57-14ca-4771-8858-ca6ee86c6400}"
all_users = true

[tool.cxfreeze.bdist_msi.summary_data]
author = "Timeseer.AI"
comments = "Kukur"
